- silence some warning spam in `textures.h`

+++ src/common/textures/textures.h
@@ -320,7 +320,7 @@ public:
 	void NeedUpdate() { bNeedsUpdate = true; }
 	void SetUpdated(bool rendertype) { bNeedsUpdate = false; bFirstUpdate = false; bLastUpdateType = rendertype; }
 
-	void SetAspectRatio(double aspectScale, bool useTextureRatio) { aspectRatio = aspectScale * (useTextureRatio? ((float)Width / Height) : 1); }
+	void SetAspectRatio(double aspectScale, bool useTextureRatio) { aspectRatio = (float)aspectScale * (useTextureRatio? ((float)Width / Height) : 1); }
 
 protected:
